+++ tests/upstream_alignment_test.py
@@ -24,15 +24,20 @@ class UpstreamAlignmentTest(TestCase):
         'mj-breakpoint',
         'mj-title',
         'mj-style',
+        'mj-accordion',
         'mj-attributes',
+        'mj-html-attributes',
         'mj-column-with-attributes',
         'mj-group',
+        'mj-hero-fixed',
+        'mj-hero-fluid',
         'mj-button-with-width',
         'mj-text-with-tail-text',
         'mj-table',
         'mj-head-with-comment',
         'mj-image-with-empty-alt-attribute',
         'mj-image-with-href',
+        'mj-section-with-full-width',
         'mj-section-with-mj-class',
         'mj-section-with-background-url',
         'mj-font',
@@ -44,6 +49,7 @@ class UpstreamAlignmentTest(TestCase):
         'mj-raw-with-tags',
         'mj-raw-head',
         'mj-raw-head-with-tags',
+        'mj-social',
         'mj-spacer',
         'mj-wrapper',
     )
@@ -110,6 +116,25 @@ class UpstreamAlignmentTest(TestCase):
 
         assert_same_html(str(expected_soup), str(actual_soup), verbose=True)
 
+    # The dynamically generated carousel ID prevents us from just using
+    # test_ensure_same_html to test mj-carousel
+    def test_mj_carousel(self):
+        test_id = 'mj-carousel'
+        expected_html = load_expected_html(test_id)
+        with get_mjml_fp(test_id) as mjml_fp:
+            result = mjml_to_html(mjml_fp)
+
+        assert not result.errors
+        expected_soup = BeautifulSoup(expected_html, 'html.parser')
+        actual_soup = BeautifulSoup(result.html, 'html.parser')
+
+        # This ID is randomly generated, so we need to manually replace it.
+        expected_carousel_id = expected_soup.find(attrs={'class': 'mj-carousel-radio'})['name'].replace('mj-carousel-radio-', '')
+        actual_carousel_id = actual_soup.find(attrs={'class': 'mj-carousel-radio'})['name'].replace('mj-carousel-radio-', '')
+        actual_html = str(actual_soup).replace(actual_carousel_id, expected_carousel_id)
+
+        assert_same_html(str(expected_soup), actual_html, verbose=True)
+
     # htmlcompare is currently unable to detect these kind of
     # whitespace differences.
     def test_keep_whitespace_before_tag(self):
@@ -125,5 +150,4 @@ class UpstreamAlignmentTest(TestCase):
         assert (expected_text == actual_text)
         actual_html = (body_actual.select('table > tr > td > div')[0]).renderContents()
         assert (b'foo <b>bar</b>.' == actual_html)
-